requirements.txt
python-telegram-bot==20.8
requests==2.32.3
aiohttp==3.12.0
python-dotenv==1.1.0